+++ src/main/kotlin/app/revanced/patches/myfitnesspal/ads/fingerprints/IsPremiumUseCaseImplFingerprint.kt
@@ -1,0 +1,11 @@
+package app.revanced.patches.myfitnesspal.ads.fingerprints
+
+import app.revanced.patcher.fingerprint.MethodFingerprint
+import com.android.tools.smali.dexlib2.AccessFlags
+
+object IsPremiumUseCaseImplFingerprint : MethodFingerprint(
+    accessFlags = AccessFlags.PUBLIC.value,
+    customFingerprint = { methodDef, classDef ->
+        classDef.type.endsWith("myfitnesspal/libs/ads/usecases/IsPremiumUseCaseImpl;") && methodDef.name == "isPremium"
+    }
+)
